+++ packages/react/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @versa/react
 
+## 1.16.0
+
+### Minor Changes
+
+- update for schema version 2.3.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @versa/belt@1.16.0
+  - @versa/pdfgen@1.16.0
+  - @versa/schema@2.3.0
+
 ## 1.15.1
 
 ### Patch Changes