+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @versa/schema
 
+## 2.3.0
+
+### Minor Changes
+
+- update for schema version 2.3.0
+
 ## 2.2.0
 
 ### Minor Changes